Updated packaging strategy
