+++ Android/app/src/main/AndroidManifest.xml
@@ -17,17 +17,17 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/Theme.AndroidGUI">
-        <activity android:name=".inventory.UpdateQuanActivity">
+        <activity android:name=".InventoryActivities.UpdateQuanActivity">
         </activity>
-        <activity android:name=".inventory.UpdateFreshnessActivity">
+        <activity android:name=".InventoryActivities.UpdateFreshnessActivity">
         </activity>
-        <activity android:name=".inventory.AddInventoryActivity">
+        <activity android:name=".InventoryActivities.AddInventoryActivity">
         </activity>
-        <activity android:name=".CheckInventoryActivity">
+        <activity android:name=".InventoryActivities.CheckInventoryActivity">
         </activity>
-        <activity android:name=".inventory.InventoryStarterActivity">
+        <activity android:name=".InventoryActivities.InventoryStarterActivity">
         </activity>
-        <activity android:name=".EnrollUserActivity">
+        <activity android:name=".UserActivities.EnrollUserActivity">
         </activity>
         <activity
                 android:name=".ManagerActivities.SelectEditOrDeleteActivity"
@@ -50,7 +50,7 @@
                 android:exported="true">
         </activity>
         <activity
-                android:name=".KitchenActivity"
+                android:name=".KitchenActivities.KitchenActivity"
                 android:exported="true">
         </activity>
         <!--
@@ -65,44 +65,44 @@
                 android:name="com.google.android.geo.API_KEY"
                 android:value="null"/> <!-- Need change previous line to real "value" -->
         <activity
-                android:name=".MapsActivity"
+                android:name=".DeliveryStaffActivities.MapsActivity"
                 android:label="@string/title_activity_maps">
         </activity>
-        <activity android:name=".ServeDishActivity">
+        <activity android:name=".ServingStaffActivities.ServeDishActivity">
         </activity>
         <activity
-                android:name=".DeliverOrderActivity"
+                android:name=".DeliveryStaffActivities.DeliverOrderActivity"
                 android:label="@string/title_activity_deliver_order"
                 android:theme="@style/Theme.AndroidGUI.NoActionBar">
         </activity>
         <activity
-                android:name=".LoginActivity"
+                android:name=".UserActivities.LoginActivity"
                 android:exported="true">
         </activity>
 
-        <activity android:name=".placeorder.SelectDineInTakeOutActivity"
+        <activity android:name=".UserActivities.SelectDineInTakeOutActivity"
                   android:exported="true">
         </activity>
-        <activity android:name=".placeorder.EnterLocationActivity"
+        <activity android:name=".UserActivities.EnterLocationActivity"
                   android:exported="true">
         </activity>
         <activity
-                android:name=".placeorder.PlaceOrderActivity"
+                android:name=".UserActivities.PlaceOrderActivity"
                 android:exported="true">
         </activity>
-        <activity android:name=".placeorder.OrderSuccessfullyPlacedActivity"
+        <activity android:name=".UserActivities.OrderSuccessfullyPlacedActivity"
         android:exported="true">
     </activity>
         <activity
-                android:name=".SeeMenuActivity">
+                android:name=".UserActivities.SeeMenuActivity">
           </activity>
-        <activity android:name=".placeorder.EditOrderActivity"
+        <activity android:name=".UserActivities.EditOrderActivity"
                 android:exported="true">
         </activity>
-        <activity android:name=".CustomerPickActionActivity">
+        <activity android:name=".UserActivities.CustomerPickActionActivity">
         </activity>
-        <activity android:name=".ServingStaffPickActionActivity"/>
-        <activity android:name=".DeliveryStaffPickActionActivity"/>
+        <activity android:name=".ServingStaffActivities.ServingStaffPickActionActivity"/>
+        <activity android:name=".DeliveryStaffActivities.DeliveryStaffPickActionActivity"/>
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
