+++ Android/app/src/main/java/com/example/androidgui/DeliveryStaffActivities/DeliverOrderActivity.java
@@ -1,4 +1,4 @@
-package com.example.androidgui;
+package com.example.androidgui.DeliveryStaffActivities;
 
 import android.content.Intent;
 import android.view.View;
@@ -6,9 +6,10 @@ import android.widget.TextView;
 import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+import com.example.androidgui.R;
 import controller.staffSystem.StaffController;
 
-public class ServeDishActivity extends AppCompatActivity {
+public class DeliverOrderActivity extends AppCompatActivity {
     private String id;
     private String mode;
     private StaffController controller;
@@ -17,44 +18,46 @@ public class ServeDishActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         controller = new StaffController();
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_serve_dish);
+        setContentView(R.layout.activity_deliver_order);
+        // Get id for method calls
         Bundle b = getIntent().getExtras();
         if(b != null) {
             id = b.getString("id");
             mode = b.getString("action");
         }
-        // Get next dish to be delivered
-        if (mode != null && mode.equals("GET_NEXT")) {
+        // Get next order to be delivered
+        if (mode.equals("GET_NEXT")) {
             Toast toast;
             try {
                 controller.getNext(this.id);
             } catch (Exception e) {
-                if (e.getMessage() != null && !e.getMessage().equals("Already has one dish in hands")) {
+                if (e.getMessage() != null && !e.getMessage().equals("Already has one order in hands")) {
                     toast = Toast.makeText(getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT);
                     toast.show();
                 }
                 goBackPickAction();
             }
         }
-        // Display current dish to be delivered
-        TextView currentDish = findViewById(R.id.CurrentDish);
+        // Display current order to be delivered
+        TextView currentOrder = findViewById(R.id.CurrentOrder);
         try {
-            currentDish.setText(controller.displayCurrent(this.id));
+            currentOrder.setText(controller.displayCurrent(this.id));
         } catch (Exception e) {
             exceptionHandler(e);
         }
     }
+
     /**
-     * When select to complete the dish in hand, try to call completeCurrent
+     * When select to finish the order in hand, try to call completeCurrent
      */
-    public void selectCompleteDish(View v) {
+    public void selectFinishOrder(View v) {
         try {
             controller.completeCurrent(this.id);
         } catch (Exception e) {
             exceptionHandler(e);
         }
         // After finishing the current order, first send a message saying that, then return to select action page
-        Toast toast = Toast.makeText(getApplicationContext(), R.string.serve_dish_complete, Toast.LENGTH_SHORT);
+        Toast toast = Toast.makeText(getApplicationContext(), R.string.deliver_order_complete, Toast.LENGTH_SHORT);
         toast.show();
         goBackPickAction();
     }
@@ -73,11 +76,11 @@ public class ServeDishActivity extends AppCompatActivity {
     private void exceptionHandler(Exception e) {
         // When exception, throw exception as toast then back to menu
         Toast toast;
-        if (e.getMessage() != null && e.getMessage().equals("Already has one dish in hands")) {
+        if (e.getMessage() != null && e.getMessage().equals("Already has one order in hands")) {
             toast = Toast.makeText(getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT);
             toast.show();
-        } else if (e.getMessage() != null && e.getMessage().equals("No current dish to be displayed")) {
-            toast = Toast.makeText(getApplicationContext(), R.string.no_current_dish, Toast.LENGTH_SHORT);
+        } else if (e.getMessage() != null && e.getMessage().equals("No current order to be displayed")) {
+            toast = Toast.makeText(getApplicationContext(), R.string.no_current_order, Toast.LENGTH_SHORT);
             toast.show();
             // Jump back to pick action page
             goBackPickAction();
@@ -93,11 +96,12 @@ public class ServeDishActivity extends AppCompatActivity {
      * Helper method for go back to pick action screen
      */
     private void goBackPickAction() {
-        Intent intent = new Intent(ServeDishActivity.this, ServingStaffPickActionActivity.class);
+        Intent intent = new Intent(DeliverOrderActivity.this, DeliveryStaffPickActionActivity.class);
         Bundle b;
         b = new Bundle();
         b.putString("id", this.id); //Your id
         intent.putExtras(b); //Put your id to next activity
         startActivity(intent);
     }
+
 }
