+++ Android/app/src/main/java/com/example/androidgui/InventoryActivities/CheckInventoryActivity.java
@@ -1,4 +1,4 @@
-package com.example.androidgui;
+package com.example.androidgui.InventoryActivities;
 
 import android.content.Intent;
 import android.view.View;
@@ -7,7 +7,8 @@ import android.widget.EditText;
 import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import com.example.androidgui.inventory.PresentInventoryActivity;
+//import com.example.androidgui.inventory.PresentInventoryActivity;
+import com.example.androidgui.R;
 import controller.inventorySystem.InventoryManager;
 
 
@@ -28,15 +29,15 @@ public class CheckInventoryActivity extends AppCompatActivity implements View.On
 
     @Override
     public void onClick(View v) {
-        String iname = name.getText().toString();
-        String info = im.getInfo(iname);
-        boolean wrongname = info.equals("Invalid name");
+        String iName = name.getText().toString();
+        String info = im.getInfo(iName);
+        boolean wrongName = info.equals("Invalid name");
         String fail = "Invalid name";
-        if (wrongname) {
+        if (wrongName) {
             Toast.makeText(CheckInventoryActivity.this,fail,Toast.LENGTH_SHORT).show();
         }else {
             Intent intent = new Intent(CheckInventoryActivity.this, PresentInventoryActivity.class);
-            intent.putExtra("showdata",info);
+            intent.putExtra("showData",info);
             startActivity(intent);
         }
     }
