+++ Android/app/src/main/java/com/example/androidgui/InventoryActivities/InventoryStarterActivity.java
@@ -1,10 +1,9 @@
-package com.example.androidgui.inventory;
+package com.example.androidgui.InventoryActivities;
 
 import android.content.Intent;
 import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import com.example.androidgui.CheckInventoryActivity;
 import com.example.androidgui.R;
 
 public class InventoryStarterActivity extends AppCompatActivity {
