+++ Android/app/src/main/java/com/example/androidgui/ManagerActivities/ManagerPickActionActivity.java
@@ -6,7 +6,7 @@ import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
 import com.example.androidgui.MainActivity;
 import com.example.androidgui.R;
-import com.example.androidgui.SeeMenuActivity;
+import com.example.androidgui.UserActivities.SeeMenuActivity;
 
 /**
  * Activity class for the manager the pick the activities he or she wants to do.
