+++ Android/app/src/main/java/com/example/androidgui/UserActivities/EnrollUserActivity.java
@@ -1,4 +1,4 @@
-package com.example.androidgui;
+package com.example.androidgui.UserActivities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,8 @@ import android.widget.EditText;
 import android.widget.NumberPicker;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
+import com.example.androidgui.R;
+import com.example.androidgui.UserActivities.LoginActivity;
 import constant.uiMessage.EnrollUserMessage;
 import constant.uiMessage.LoginLogoutUIMessage;
 import controller.managerSystem.ManagerController;
