+++ Android/app/src/main/java/com/example/androidgui/UserActivities/OrderSuccessfullyPlacedActivity.java
@@ -1,10 +1,9 @@
-package com.example.androidgui.placeorder;
+package com.example.androidgui.UserActivities;
 
 import android.content.Intent;
 import android.view.View;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import com.example.androidgui.CustomerPickActionActivity;
 import com.example.androidgui.R;
 
 public class OrderSuccessfullyPlacedActivity extends AppCompatActivity {
