+++ Android/app/src/main/java/com/example/androidgui/UserActivities/PlaceOrderActivity.java
@@ -1,4 +1,4 @@
-package com.example.androidgui.placeorder;
+package com.example.androidgui.UserActivities;
 
 import android.content.Intent;
 import android.os.Parcelable;
