+++ Android/app/src/main/java/com/example/androidgui/UserActivities/SeeMenuActivity.java
@@ -1,9 +1,11 @@
-package com.example.androidgui;
+package com.example.androidgui.UserActivities;
 
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+import com.example.androidgui.MainActivity;
+import com.example.androidgui.R;
 
 
 public class SeeMenuActivity extends AppCompatActivity {
