+++ Android/app/src/main/java/use_case/serveDish/ServingBuffer.java
@@ -13,7 +13,7 @@ public class ServingBuffer {
     private static final Queue<Dish> servingBuffer = new ArrayDeque<>();
 
     /**
-     * The getter method for UseCase.ServingStaff to call to get the next dish to serve.
+     * The getter method for UseCase.com.example.androidgui.ServingStaff to call to get the next dish to serve.
      * @return a UseCase.DishInfo instance, which contains the table number of the dish and the dish itself.
      */
     public static Dish getNextToServe() throws Exception{
