+++ Android/app/src/main/res/layout/activity_add_inventory2.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".inventory.AddInventoryActivity">
+        tools:context=".InventoryActivities.AddInventoryActivity">
 
     <EditText
             android:layout_width="358dp"
