+++ Android/app/src/main/res/layout/activity_add_review_comment.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".AddReviewCommentActivity">
+        tools:context=".UserActivities.AddReviewCommentActivity">
 
     <com.google.android.material.textfield.TextInputEditText
             android:layout_width="0dp"
