+++ Android/app/src/main/res/layout/activity_check_inventory2.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CheckInventoryActivity">
+        tools:context=".InventoryActivities.CheckInventoryActivity">
 
     <Button
             android:text="Search"
