+++ Android/app/src/main/res/layout/activity_customer_pick_action.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CustomerPickActionActivity">
+        tools:context=".UserActivities.CustomerPickActionActivity">
 
     <TextView
             android:text="@string/welcome_customer"
