+++ Android/app/src/main/res/layout/activity_deliver_order.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DeliverOrderActivity">
+    tools:context=".DeliveryStaffActivities.DeliverOrderActivity">
 
   <com.google.android.gms.maps.MapView
           android:layout_width="358dp"
