+++ Android/app/src/main/res/layout/activity_edit_order.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".placeorder.EditOrderActivity">
+        tools:context=".UserActivities.EditOrderActivity">
     <NumberPicker android:layout_width="270dp" android:layout_height="103dp"
                   app:layout_constraintTop_toTopOf="parent"
                   android:layout_marginTop="116dp" app:layout_constraintStart_toStartOf="parent"
