+++ Android/app/src/main/res/layout/activity_enroll_user.xml
@@ -4,7 +4,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context=".EnrollUserActivity">
+	tools:context=".UserActivities.EnrollUserActivity">
 
 	<TextView
 		android:id="@+id/textView6"
