+++ Android/app/src/main/res/layout/activity_enter_location.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".placeorder.EnterLocationActivity">
+        tools:context=".UserActivities.EnterLocationActivity">
 
     <EditText
             android:layout_width="wrap_content"
