+++ Android/app/src/main/res/layout/activity_inventory_starter.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".inventory.InventoryStarterActivity">
+        tools:context=".InventoryActivities.InventoryStarterActivity">
 
     <Button
             android:text="@string/add_inventory"
