+++ Android/app/src/main/res/layout/activity_kitchen.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".KitchenActivity">
+        tools:context=".KitchenActivities.KitchenActivity">
     <ListView
             android:layout_width="match_parent"
             android:layout_height="wrap_content" tools:layout_editor_absoluteX="1dp"
