+++ Android/app/src/main/res/layout/activity_maps.xml
@@ -5,5 +5,5 @@
           android:layout_width="match_parent"
           android:layout_height="match_parent"
           android:id="@+id/map"
-          tools:context=".MapsActivity"
+          tools:context=".DeliveryStaffActivities.MapsActivity"
           android:name="com.google.android.gms.maps.SupportMapFragment"/>
