+++ Android/app/src/main/res/layout/activity_order_successfully_placed.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".placeorder.OrderSuccessfullyPlacedActivity">
+        tools:context=".UserActivities.OrderSuccessfullyPlacedActivity">
 
     <TextView
             android:text="@string/order_placed_message"
