+++ Android/app/src/main/res/layout/activity_place_order.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".placeorder.PlaceOrderActivity">
+        tools:context=".UserActivities.PlaceOrderActivity">
     <NumberPicker android:layout_width="63dp" android:layout_height="87dp"
                   android:id="@+id/dishQuantityPicker" android:layout_marginTop="26dp"
                   app:layout_constraintTop_toBottomOf="@+id/textView2"
