+++ Android/app/src/main/res/layout/activity_present_inventory.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".inventory.PresentInventoryActivity">
+        tools:context=".InventoryActivities.PresentInventoryActivity">
 
     <TextView
             android:text="@string/textview"
