+++ Android/app/src/main/res/layout/activity_see_menu.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SeeMenuActivity">
+        tools:context=".UserActivities.SeeMenuActivity">
 
     <TextView
             android:text="@string/menu_title"
