+++ Android/app/src/main/res/layout/activity_select_dine_in_take_out.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".placeorder.SelectDineInTakeOutActivity">
+        tools:context=".UserActivities.SelectDineInTakeOutActivity">
     <NumberPicker android:layout_width="160dp" android:layout_height="66dp"
                   app:layout_constraintTop_toTopOf="parent"
                   android:layout_marginTop="284dp" app:layout_constraintStart_toStartOf="parent"
