+++ Android/app/src/main/res/layout/activity_serve_dish.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ServeDishActivity">
+        tools:context=".ServingStaffActivities.ServeDishActivity">
     <TextView
             android:text="@string/textview"
             android:layout_width="400dp"
