+++ Android/app/src/main/res/layout/activity_update_freshness2.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".inventory.UpdateFreshnessActivity">
+        tools:context=".InventoryActivities.UpdateFreshnessActivity">
 
     <EditText
             android:layout_width="wrap_content"
