+++ Android/app/src/main/res/layout/activity_update_quan2.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".inventory.UpdateQuanActivity">
+        tools:context=".InventoryActivities.UpdateQuanActivity">
 
     <Button
             android:text="@string/update"
